Renamed networking index file to api. (Leaving room for websocket)
